requirements.txt
scipy==1.5.2
pyparsing==2.4.7
matplotlib==3.3.3
markdown==3.3.4
pytorch-lightning==0.7.5
scikit-image==0.16.2
nibabel==2.5.1
opencv-python==4.1.2.30
pyyaml==5.4.1
h5py==2.10.0
batchgenerators==0.19.5
pydicom==1.4.2
SimpleITK==1.2.4
pandas==1.1
pycodestyle==2.7.0
flake8==3.9.2
autopep8==1.5.7
black==22.3.0
termcolor==1.1.0
numpy==1.19.5
scikit-learn==0.23.0
setuptools==45.3.0
openslide-python==1.1.1
kornia==0.6.6
dicompyler-core==0.5.5
shapely==1.7.0
ansible-vault==1.2.0
pyarmor==6.4.0
pyradiomics==3.0.1
python-magic==0.4.24
timm==0.5.4
medpy==0.4.0
pre-commit==2.17.0
monai==0.8.1
ACSConv==0.1.1
pytest==7.0.1
einops==0.4.1
gnureadline==8.0.0